refactor: 엔티티 객체 필드명 개선

+++ src/main/java/com/dgmoonlabs/psythinktank/domain/circular/repository/CircularRepository.java
@@ -3,5 +3,5 @@ package com.dgmoonlabs.psythinktank.domain.circular.repository;
 import com.dgmoonlabs.psythinktank.domain.circular.model.Circular;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CircularRepository extends JpaRepository<Circular, Integer> {
+public interface CircularRepository extends JpaRepository<Circular, Long> {
 }
